fix problem with filter markers, add propTypes to User component, add conditional statment that to not delete first user(admin), change isNavSelect action name to isPanelSelect, disable Filter link at markerCreator component

+++ src/controllers/user/deleteAccount.ts
@@ -5,6 +5,9 @@ import { User } from "../../entity/User";
 export async function deleteAccount(request: Request, response: Response) {
   const userRepository = getManager().getRepository(User);
   const userToDelete = await userRepository.findOne(request.params.id);
-  userRepository.remove(userToDelete);
+
+  if (userToDelete.id !== 1) {
+    userRepository.remove(userToDelete);
+  }
   return response.json(userToDelete);
 }
